+++ packages/terra-embedded-content-consumer/tests/wdio/embedded-content-consumer-spec.js
@@ -1,18 +1,18 @@
 Terra.describeViewports('Embedded Content Consumer', ['tiny', 'large'], () => {
   it('displays default embedded content consumer', () => {
-    browser.url('/raw/tests/terra-embedded-content-consumer/embedded-content-consumer/consumers/basic-consumer');
+    browser.url('/raw/tests/cerner-terra-framework-docs/embedded-content-consumer/consumers/basic-consumer');
     Terra.validates.element('default');
   });
 
   describe('custom-event', () => {
     it('displays with custom events', () => {
-      browser.url('/raw/tests/terra-embedded-content-consumer/embedded-content-consumer/consumers/custom-event-consumer');
+      browser.url('/raw/tests/cerner-terra-framework-docs/embedded-content-consumer/consumers/custom-event-consumer');
       $('#CustomEvent').waitForDisplayed();
       Terra.validates.element('custom event');
     });
 
     it('Provider triggers EventA message', () => {
-      const myFrame = $('iframe[src="/raw/provider/terra-embedded-content-consumer/embedded-content-consumer/providers/custom-event-provider"]');
+      const myFrame = $('iframe[src="/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/custom-event-provider"]');
       browser.switchToFrame(myFrame);
 
       $('#EventA').click();
@@ -24,13 +24,13 @@ Terra.describeViewports('Embedded Content Consumer', ['tiny', 'large'], () => {
 
   describe('custom-events', () => {
     it('displays with custom event', () => {
-      browser.url('/raw/tests/terra-embedded-content-consumer/embedded-content-consumer/consumers/custom-events-consumer');
+      browser.url('/raw/tests/cerner-terra-framework-docs/embedded-content-consumer/consumers/custom-events-consumer');
       $('#CustomEvents').waitForDisplayed();
       Terra.validates.element('custom events');
     });
 
     it('Provider triggers EventA message', () => {
-      const myFrame = $('iframe[src="/raw/provider/terra-embedded-content-consumer/embedded-content-consumer/providers/custom-events-provider"]');
+      const myFrame = $('iframe[src="/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/custom-events-provider"]');
       browser.switchToFrame(myFrame);
 
       $('#EventA').click();
@@ -40,7 +40,7 @@ Terra.describeViewports('Embedded Content Consumer', ['tiny', 'large'], () => {
     });
 
     it('successfully replied with EventA message', () => {
-      const myFrame = $('iframe[src="/raw/provider/terra-embedded-content-consumer/embedded-content-consumer/providers/custom-events-provider"]');
+      const myFrame = $('iframe[src="/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/custom-events-provider"]');
       browser.switchToFrame(myFrame);
 
       expect($('#embedded-content-consumer-reply')).toHaveTextContaining('eventA');
@@ -48,7 +48,7 @@ Terra.describeViewports('Embedded Content Consumer', ['tiny', 'large'], () => {
     });
 
     it('Provider triggers EventB message', () => {
-      const myFrame = $('iframe[src="/raw/provider/terra-embedded-content-consumer/embedded-content-consumer/providers/custom-events-provider"]');
+      const myFrame = $('iframe[src="/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/custom-events-provider"]');
       browser.switchToFrame(myFrame);
 
       $('#EventB').click();
@@ -58,7 +58,7 @@ Terra.describeViewports('Embedded Content Consumer', ['tiny', 'large'], () => {
     });
 
     it('successfully replied with EventB message', () => {
-      const myFrame = $('iframe[src="/raw/provider/terra-embedded-content-consumer/embedded-content-consumer/providers/custom-events-provider"]');
+      const myFrame = $('iframe[src="/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/custom-events-provider"]');
       browser.switchToFrame(myFrame);
 
       expect($('#embedded-content-consumer-reply')).toHaveTextContaining('eventB');
@@ -67,11 +67,11 @@ Terra.describeViewports('Embedded Content Consumer', ['tiny', 'large'], () => {
 
   describe('data-status', () => {
     it('has mounted, launched, and authorized elements', () => {
-      browser.url('/raw/tests/terra-embedded-content-consumer/embedded-content-consumer/consumers/data-status-consumer');
+      browser.url('/raw/tests/cerner-terra-framework-docs/embedded-content-consumer/consumers/data-status-consumer');
       const timeout = browser.options.waitforTimeout + 5000;
-      $('iframe[src="/raw/provider/terra-embedded-content-consumer/embedded-content-consumer/providers/data-status-provider"]').waitForDisplayed({ timeout });
+      $('iframe[src="/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/data-status-provider"]').waitForDisplayed({ timeout });
 
-      const myFrame = $('iframe[src="/raw/provider/terra-embedded-content-consumer/embedded-content-consumer/providers/data-status-provider"]');
+      const myFrame = $('iframe[src="/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/data-status-provider"]');
       browser.switchToFrame(myFrame);
 
       expect($('#Mounted').isExisting());