+++ packages/terra-framework-docs/src/terra-dev-site/test/slide-panel/SlidePanelStart.test.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import SlidePanel, { SlidePanelPositions } from 'terra-slide-panel';
+import SlidePanelMainContent from './common/SlidePanelMainContent';
+import SlidePanelPanelContent from './common/SlidePanelPanelContent';
+
+const SlidePanelStart = () => (
+  <SlidePanel
+    mainContent={<SlidePanelMainContent />}
+    panelContent={<SlidePanelPanelContent />}
+    panelBehavior="overlay"
+    panelPosition={SlidePanelPositions.START}
+    panelSize="small"
+    isOpen
+  />
+);
+
+export default SlidePanelStart;